milestones/milestone4/orm_project/requirements.txt
mysql-connector-python==8.0.27
pytest==6.2.5
#pysqlite3==0.4.6